fix: convert BigDecimal to Double for OpenAI temperature parameter
- Fix compilation error in AIAgentService.enhanceMessage method
- Convert agent.getTemperature() from BigDecimal to Double using doubleValue()
- Ensures compatibility with OpenAI service that expects Double temperature

+++ api/src/main/java/com/ruby/rubia_server/core/service/AIAgentService.java
@@ -232,7 +232,7 @@ public class AIAgentService {
             return openAIService.enhanceTemplate(
                 enhancementPrompt,
                 agent.getAiModel().getName(),
-                agent.getTemperature(),
+                agent.getTemperature().doubleValue(),
                 agent.getMaxResponseLength()
             );
         } catch (Exception e) {
